Separate mutable table into its own file. Add local storage
